# docs/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bitedash API",
    "version": "1.0.0",
    "description": "Minimal OpenAPI spec for Bitedash backend endpoints used by admin and mobile clients."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Local development"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/LoginRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthResponse"}}}
          },
          "401": {"description": "Invalid credentials"}
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register user",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"201": {"description": "Created"}}
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "requestBody": {"required": true,"content": {"application/json": {"schema": {"type": "object","properties":{"refresh_token":{"type":"string"}}}}}},
        "responses": {"200": {"description": "New access token","content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthResponse"}}}}}
      }
    },
    "/auth/profile": {
      "get": {
        "summary": "Get profile",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "User profile","content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}}
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {"type": "string"},
          "password": {"type": "string"}
        },
        "required": ["email","password"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "access_token": {"type": "string"},
          "refresh_token": {"type": "string"},
          "expires_in": {"type": "integer"}
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "email": {"type": "string"},
          "name": {"type": "string"}
        }
      }
    }
  }
}